requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
alembic>=1.13.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.1.0

